refactor: #9 DeliveryMapper 사용
