+++ delivery-service/src/main/java/com/pickple/delivery/application/dto/DeliveryCreateResponseDto.java
@@ -1,6 +1,5 @@
 package com.pickple.delivery.application.dto;
 
-import com.pickple.delivery.domain.model.Delivery;
 import com.pickple.delivery.domain.model.enums.DeliveryStatus;
 import java.util.UUID;
 import lombok.Builder;
@@ -21,17 +20,5 @@ public class DeliveryCreateResponseDto {
     private String recipientAddress;
 
     private String recipientContact;
-
-    public static DeliveryCreateResponseDto fromEntity(Delivery delivery) {
-        return DeliveryCreateResponseDto.builder()
-                .deliveryId(delivery.getDeliveryId())
-                .orderId(delivery.getOrderId())
-                .deliveryStatus(delivery.getDeliveryStatus())
-                .deliveryRequirement(delivery.getDeliveryRequirement())
-                .recipientName(delivery.getRecipientName())
-                .recipientAddress(delivery.getRecipientAddress())
-                .recipientContact(delivery.getRecipientContact())
-                .build();
-    }
 
 }
